Final commit
Contains the most recent synthesis from Demo Day that was
living in Horizon. system_top has controllers built in,
but the GPU wasn't latching instructions/memory wasn't
sending it the right instructions and so we don't get to see
buggy Pong on the screen.
It is set to run the PSX bios right now, not buggy Pong